requirements.txt
GitPython
pydantic
openai
python-dotenv